Moved About.js to sections/
